Almost finishing the rename client (#401)
* almost finishing the rename

* Fixing references
