+++ client/src/pages/signal-generator/examples/bpsk-tone-noise.ts
