+++ client/src/pages/signal-generator/examples/direct-sequence-spread-spectrum.ts
