+++ client/src/pages/signal-generator/examples/example-list.ts
@@ -1,6 +1,6 @@
-import { bpskToneNoise } from './bpskToneNoise';
+import { bpskToneNoise } from './bpsk-tone-noise';
 import { ofdm } from './ofdm';
-import { directSequenceSpreadSpectrum } from './directSequenceSpreadSpectrum';
+import { directSequenceSpreadSpectrum } from './direct-sequence-spread-spectrum';
 
 // Add yours to the end of this list, and remember to import it above
 export const examplesList = [bpskToneNoise, ofdm, directSequenceSpreadSpectrum];
