+++ client/src/pages/signal-generator/signal-generator.tsx
@@ -5,9 +5,9 @@ import React, { useEffect, useState, useCallback } from 'react';
 import { vscodeDark } from '@uiw/codemirror-theme-vscode';
 import CodeMirror from '@uiw/react-codemirror';
 import { python } from '@codemirror/lang-python';
-import { examplesList } from '@/pages/signal-generator/examples/exampleList';
+import { examplesList } from '@/pages/signal-generator/examples/example-list';
 
-export default function SignalGenerator(props) {
+export function SignalGenerator(props) {
   const [state, setState] = useState({
     freqPlotSnippet: `\
 X = 10*np.log10(np.abs(np.fft.fftshift(np.fft.fft(x)))**2)
